analyzer/requirements.txt
google-generativeai>=0.3.0
firebase-admin>=6.4.0
Pillow>=10.2.0
aiohttp>=3.9.3
tqdm>=4.66.2
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.5
